new argument: tplqmin
tplqmin is an additional argument for serval. It allows to choose the minimum template quality factor. Template knots with lower quality factor are flagged and not used to compute radial velocities. The template quality factor is the ratio of good pixels to all pixels in the template generation and computed for each template knot.

tplqmin replaces the serval internal variable rmin.