Update cspline.py
Adds option to compute the AIC and BIC information criteria using the effective degrees of freedom. Computing the AIC or BIC allows to optimize the pspline penalty. The effective degrees of freedom replace the classic degrees of freedom in the computation of the AIC and BIC for the pspline (Eilers & Marx, 1996).