Update inst_HPF.py
The updated instrument file can handle reduced HPF spectra from the Goldilocks and the PSU pipelines. Deblazing was added for the Goldilocks spectra. The script further contains functions for scaling and subtraction of the HPF sky channel from the HPF science channel. The sky subtraction is disabled, because it did not improve RV precision compared to applying a sky mask.